Fix verify_signature() API
Fix compiler warning:
pam_pkcs11.c:729:62: warning: passing argument 5 of ‘verify_signature’ from incompatible pointer type [-Wincompatible-pointer-types]
              random_value, sizeof(random_value), &signature, &signature_length);
                                                              ^
In file included from pam_pkcs11.c:45:0:
../common/cert_vfy.h:81:20: note: expected ‘int *’ but argument is of type ‘long unsigned int *’
 CERTVFY_EXTERN int verify_signature(X509 * x509, unsigned char *data, int data_length, unsigned char **signature, int *signature_length);
                    ^~~~~~~~~~~~~~~~

+++ src/common/cert_vfy.c
@@ -42,7 +42,7 @@ int verify_certificate(X509 * x509, cert_policy *policy)
 }
 
 int verify_signature(X509 * x509, unsigned char *data, int data_length,
-                     unsigned char **signature, int *signature_length)
+                     unsigned char **signature, unsigned long *signature_length)
 {
 
   SECKEYPublicKey *key;
@@ -499,7 +499,7 @@ int verify_certificate(X509 * x509, cert_policy *policy)
 }
 
 int verify_signature(X509 * x509, unsigned char *data, int data_length,
-                     unsigned char **signature, int *signature_length)
+                     unsigned char **signature, unsigned long *signature_length)
 {
   int rv;
   EVP_PKEY *pubkey;
